requirements.txt
# Core PDF Processing
pymupdf4llm>=0.0.1
pymupdf>=1.23.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# Data Validation and Models
pydantic>=2.0.0

# CLI
click>=8.1.0

# Search and Embeddings
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
sentence-transformers>=2.2.0

# Configuration and Logging
python-dotenv>=1.0.0
loguru>=0.7.0

# LLM Providers (optional, install as needed)
# For Ollama: pip install requests (already included)
# For HuggingFace: pip install transformers torch
# For Anthropic: pip install anthropic
# For OpenAI: pip install openai
requests>=2.31.0  # Required for Ollama provider
